guards folder relocated to infrastructure folder.
