+++ Frontend/InventoryApp/src/app/dashboard/dashboard-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 // Guards
-import { AuthGuard } from '../shared/guards/auth-guard.service';
+import { AuthGuard } from '../infrastructure/guards/auth-guard.service';
 // Components
 import { DashboardComponent } from './dashboard.component';
 import { ProfileComponent } from './common/profile/profile.component';
