+++ Frontend/InventoryApp/src/app/infrastructure/guards/auth-guard.service.ts
